IN-103 / refactor: refactor Avatar Dialog
