+++ src/common/color-palette/ColorPalette.mdx
@@ -1,5 +1,3 @@
-[//]: # "import {  } from '@storybook/addon-docs'"
-
 import { ColorPalette, ColorItem, Meta } from '@storybook/blocks'
 
 <Meta title='Design/Colors' />
